+++ PluginQuery/src/septogeddon/pluginquery/netty/QueryEncoder.java
@@ -9,7 +9,7 @@ public class QueryEncoder extends MessageToByteEncoder<QueryMessage> {
 
 	@Override
 	protected void encode(ChannelHandlerContext arg0, QueryMessage arg1, ByteBuf arg2) throws Exception {
-		arg2.writeInt(arg1.getChannel().length());
+		arg2.writeByte(arg1.getChannel().length());
 		arg2.writeBytes(arg1.getChannel().getBytes());
 		arg2.writeInt(arg1.getMessage().length);
 		arg2.writeBytes(arg1.getMessage());
